libcanopend-py/pyproject.toml
[build-system]
requires = ["setuptools", "setuptools_scm"]
build-backend = "setuptools.build_meta"

[project]
name = "oresat-libcanopend"
description = "library to interface to oresat-canopend"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "GPL-3.0"}
classifiers = [
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "pyzmq",
]
dynamic = ["version"]

[project.optional-dependencies]
dev = [
    "build",
    "setuptools_scm",
    "ruff",
]

[project.urls]
homepage = "https://www.oresat.org"
download = "https://github.com/oresat/oresat-canopend/releases"
source = "https://github.com/oresat/oresat-canopend"
tracker = "https://github.com/oresat/oresat-canopend/issues"

[tool.ruff]
line-length = 100
indent-width = 4
target-version = "py39"

[tool.ruff.format]
quote-style = "double"
